feat(buildCSPHeaders): allow merging of default directives
Allow the setting of `contentSecurityPolicy.mergeDefaultDirectives`
which allows CSP directives defined in config to be additive to the
default directives defined in `next-safe`.

Also allow interpretation of string directives, e.g. `"'self' data:"`
and split these in order to merge and de-duplicate against defaults.

In addition, deal with the possibility of duplicates when merging in
development default CSP directives.

Update documentation to convey new option and usage.